refactor: create load-constants helper
